+++ tests/Unit/EmogrifierTest.php
@@ -371,6 +371,24 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @test
+     *
+     * @param string $html
+     *
+     * @dataProvider wbrTagDataProvider
+     */
+    public function emogrifyAfterRemoveUnprocessableHtmlTagKeepsWbrTag($html)
+    {
+        $numberOfWbrTags = \substr_count($html, '<wbr');
+        $this->subject->setHtml($html);
+
+        $this->subject->removeUnprocessableHtmlTag('wbr');
+        $result = $this->subject->emogrify();
+
+        static::assertSame($numberOfWbrTags, \substr_count($result, '<wbr'));
+    }
+
+    /**
+     * @test
      */
     public function addUnprocessableTagRemovesEmptyTag()
     {
@@ -2299,21 +2317,190 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * @return string[][]
+     */
+    public function xmlSelfClosingTagDataProvider()
+    {
+        return [
+            '<br>' => ['<br/>', 'br'],
+            '<wbr>' => ['foo<wbr/>bar', 'wbr'],
+            '<embed>' => [
+                '<embed type="video/mp4" src="https://example.com/flower.mp4" width="250" height="200"/>',
+                'embed',
+            ],
+            '<picture> with <source> and <img>' => [
+                '<picture><source srcset="https://example.com/flower-800x600.jpeg" media="(min-width: 600px)"/>'
+                    . '<img src="https://example.com/flower-400x300.jpeg"/></picture>',
+                'source',
+            ],
+            '<video> with <track>' => [
+                '<video controls width="250" src="https://example.com/flower.mp4">'
+                    . '<track default kind="captions" srclang="en" src="https://example.com/flower.vtt"/></video>',
+                'track',
+            ],
+        ];
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function nonXmlSelfClosingTagDataProvider()
+    {
+        return \array_map(
+            function (array $dataset) {
+                $dataset[0] = \str_replace('/>', '>', $dataset[0]);
+                return $dataset;
+            },
+            $this->xmlSelfClosingTagDataProvider()
+        );
+    }
+
+    /**
+     * @return string[][] Each dataset has three elements in the following order:
+     *         - HTML with non-XML self-closing tags (e.g. "...<br>...");
+     *         - The name of a self-closing tag contained in the HTML (e.g. "br").
+     *         - The equivalent HTML with XML self-closing tags (e.g. "...<br/>...");
+     */
+    public function selfClosingTagDataProvider()
+    {
+        return \array_map(
+            function (array $dataset) {
+                return [
+                    \str_replace('/>', '>', $dataset[0]),
+                    $dataset[1],
+                    $dataset[0],
+                ];
+            },
+            $this->xmlSelfClosingTagDataProvider()
+        );
+    }
+
+    /**
+     * Concatenates pairs of datasets (in a similar way to SQL `JOIN`) such that each new dataset consists of a 'row'
+     * from a left-hand-side dataset joined with a 'row' from a right-hand-side dataset.
+     *
+     * @param string[][] $leftDatasets
+     * @param string[][] $rightDatasets
+     *
+     * @return string[][] The new datasets comprise the first dataset from the left-hand side with each of the datasets
+     * from the right-hand side, and the each of the remaining datasets from the left-hand side with the first dataset
+     * from the right-hand side.
+     */
+    public static function joinDatasets(array $leftDatasets, array $rightDatasets)
+    {
+        $datasets = [];
+        $doneFirstLeft = false;
+        foreach ($leftDatasets as $leftDatasetName => $leftDataset) {
+            foreach ($rightDatasets as $rightDatasetName => $rightDataset) {
+                $datasets[$leftDatasetName . ' & ' . $rightDatasetName]
+                    = \array_merge($leftDataset, $rightDataset);
+                if ($doneFirstLeft) {
+                    // Not all combinations are required,
+                    // just all of 'right' with one of 'left' and all of 'left' with one of 'right'.
+                    break;
+                }
+            }
+            $doneFirstLeft = true;
+        }
+        return $datasets;
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function documentTypeAndSelfClosingTagDataProvider()
+    {
+        return static::joinDatasets($this->documentTypeDataProvider(), $this->selfClosingTagDataProvider());
+    }
+
+    /**
      * @test
      *
      * @param string $documentType
+     * @param string $htmlWithNonXmlSelfClosingTags
+     * @param string $tagName
+     * @param string $htmlWithXmlSelfClosingTags
      *
-     * @dataProvider documentTypeDataProvider
+     * @dataProvider documentTypeAndSelfClosingTagDataProvider
      */
-    public function emogrifyConvertsXmlSelfClosingTagsToNonXmlSelfClosingTag($documentType)
-    {
+    public function emogrifyConvertsXmlSelfClosingTagsToNonXmlSelfClosingTag(
+        $documentType,
+        $htmlWithNonXmlSelfClosingTags,
+        $tagName,
+        $htmlWithXmlSelfClosingTags
+    ) {
+        $this->subject->removeUnprocessableHtmlTag($tagName);
         $this->subject->setHtml(
-            $documentType . '<html><body><br/></body></html>'
+            $documentType . '<html><body>' . $htmlWithXmlSelfClosingTags . '</body></html>'
         );
 
         $result = $this->subject->emogrify();
 
-        static::assertContains('<br>', $result);
+        static::assertContains('<body>' . $htmlWithNonXmlSelfClosingTags . '</body>', $result);
+    }
+
+    /**
+     * @test
+     *
+     * @param string $documentType
+     * @param string $htmlWithNonXmlSelfClosingTags
+     * @param string $tagName
+     *
+     * @dataProvider documentTypeAndSelfClosingTagDataProvider
+     */
+    public function emogrifyKeepsNonXmlSelfClosingTags($documentType, $htmlWithNonXmlSelfClosingTags, $tagName)
+    {
+        $this->subject->removeUnprocessableHtmlTag($tagName);
+        $this->subject->setHtml(
+            $documentType . '<html><body>' . $htmlWithNonXmlSelfClosingTags . '</body></html>'
+        );
+
+        $result = $this->subject->emogrify();
+
+        static::assertContains('<body>' . $htmlWithNonXmlSelfClosingTags . '</body>', $result);
+    }
+
+    /**
+     * @test
+     *
+     * @param string $htmlWithNonXmlSelfClosingTags
+     * @param string $tagName
+     *
+     * @dataProvider nonXmlSelfClosingTagDataProvider
+     */
+    public function emogrifyNotAddsClosingTagForSelfClosingTags($htmlWithNonXmlSelfClosingTags, $tagName)
+    {
+        $this->subject->removeUnprocessableHtmlTag($tagName);
+        $this->subject->setHtml(
+            '<html><body>' . $htmlWithNonXmlSelfClosingTags . '</body></html>'
+        );
+
+        $result = $this->subject->emogrify();
+
+        static::assertNotContains('</' . $tagName, $result);
+    }
+
+    /**
+     * @test
+     *
+     * @param string $htmlWithNonXmlSelfClosingTags
+     * @param string $tagName
+     *
+     * @dataProvider nonXmlSelfClosingTagDataProvider
+     */
+    public function getDomDocumentVoidElementNotHasChildNodes($htmlWithNonXmlSelfClosingTags, $tagName)
+    {
+        $this->subject->setHtml(
+            // Append a 'trap' element that might become a child node if the HTML is parsed incorrectly
+            '<html><body>' . $htmlWithNonXmlSelfClosingTags . '<span>foo</span></body></html>'
+        );
+
+        $domDocument = $this->subject->getDomDocument();
+
+        $voidElements = $domDocument->getElementsByTagName($tagName);
+        foreach ($voidElements as $element) {
+            static::assertFalse($element->hasChildNodes());
+        }
     }
 
     /**
@@ -2369,6 +2556,26 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
         $result = $this->subject->emogrifyBodyContent();
 
         static::assertSame('<p></p>', $result);
+    }
+
+    /**
+     * @test
+     *
+     * @param string $htmlWithNonXmlSelfClosingTags
+     * @param string $tagName
+     *
+     * @dataProvider nonXmlSelfClosingTagDataProvider
+     */
+    public function emogrifyBodyContentNotAddsClosingTagForSelfClosingTags($htmlWithNonXmlSelfClosingTags, $tagName)
+    {
+        $this->subject->removeUnprocessableHtmlTag($tagName);
+        $this->subject->setHtml(
+            '<html><body>' . $htmlWithNonXmlSelfClosingTags . '</body></html>'
+        );
+
+        $result = $this->subject->emogrifyBodyContent();
+
+        static::assertNotContains('</' . $tagName, $result);
     }
 
     /**